exercises(allergies): improve handling of larger allergen count
An EnumSet is backed by a StaticBitSet, which is an IntegerBitSet iff
the number of allergens is <= the bit width of usize [1]:

    /// Returns the optimal static bit set type for the specified number
    /// of elements.  The returned type will perform no allocations,
    /// can be copied by value, and does not require deinitialization.
    /// Both possible implementations fulfill the same interface.
    pub fn StaticBitSet(comptime size: usize) type {
        if (size <= @bitSizeOf(usize)) {
            return IntegerBitSet(size);
        } else {
            return ArrayBitSet(usize, size);
        }
    }

[1] https://github.com/ziglang/zig/blob/8f481dfc3c4f/lib/std/bit_set.zig#L37-L47

+++ exercises/practice/allergies/.meta/example.zig
@@ -19,7 +19,7 @@ pub const Allergen = enum(u8) {
         return score & @enumToInt(allergen) != 0;
     }
 
-    pub fn initEnumSet(score: u64) EnumSet(Self) {
+    pub fn initEnumSet(score: usize) EnumSet(Self) {
         const len = @typeInfo(Self).Enum.fields.len;
         const bits = IntegerBitSet(len){ .mask = @truncate(tag_type, score) };
         return .{ .bits = bits };
